+++ src/main/java/flab/nutridiary/commom/config/S3Config.java
@@ -3,7 +3,7 @@ package flab.nutridiary.commom.config;
 import com.amazonaws.client.builder.AwsClientBuilder;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3Client;
-import flab.nutridiary.commom.file.S3FileStore;
+import flab.nutridiary.file.S3FileStore;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
