+++ src/main/java/flab/nutridiary/file/S3FileStore.java
@@ -1,4 +1,4 @@
-package flab.nutridiary.commom.file;
+package flab.nutridiary.file;
 
 
 import com.amazonaws.services.s3.AmazonS3;
